+++ http-device/task/app.js
@@ -78,7 +78,16 @@ function getDeviceType(deviceType){
 }
 
 function getSettings(deviceType){
-    return settings[deviceType] || settings['Default'];
+    if(deviceType && deviceType!==''){
+        if(settings.hasOwnProperty(deviceType)){
+            return settings[deviceType];
+        }else{
+            console.error('requested device type is not available');
+        }
+    }else{
+        return settings['Default'];
+    }
+    return undefined;
 }
 
 function getDeviceId(deviceId, settings){
@@ -257,28 +266,48 @@ app.put('/settings', function (req, res) {
     return result===true ? res.sendStatus(200) : res.status(400).send({error:{message: result.message}});
 });
 
-app.listen(3000, function () {
-    console.log('HTTP Device Plugin is now running with the following configuration:');
-    console.log("HOST=" + HOST);
-    console.log("HTTT_PORT=" + HTTP_PORT);
-    console.log("HTTP_SSL_PORT=" + HTTP_SSL_PORT);
-    console.log("TOKEN=" + TOKEN);
-    console.log("USER=" + USER);
-    console.log("PLUGIN=" + PLUGIN);
-    console.log("VERSION=" + VERSION);
-    console.log("DEVELOPMENT=" + DEVELOPMENT);
+function launchServer() {
+
+    function startServer() {
+        app.listen(3000, function () {
+            console.log('HTTP Device Plugin is now running with the following configuration:');
+            console.log("HOST=" + HOST);
+            console.log("HTTT_PORT=" + HTTP_PORT);
+            console.log("HTTP_SSL_PORT=" + HTTP_SSL_PORT);
+            console.log("TOKEN=" + TOKEN);
+            console.log("USER=" + USER);
+            console.log("PLUGIN=" + PLUGIN);
+            console.log("VERSION=" + VERSION);
+            console.log("DEVELOPMENT=" + DEVELOPMENT);
+        });
+    };
 
     thinger.getPluginProperty('settings').then(function (response) {
         settings = response.data.value;
         console.log("read existing settings:",JSON.stringify(settings));
         compileCallbacks();
-    }).catch(function (error) {
-        console.error("plugin settings not available");
-        settings = {
-            'Default' : {
-                auto_provision_resources : true
-            }
-        };
-    });
 
-});
+        startServer();
+
+    }).catch(function (error) {
+
+        if ( error.response && error.response.status === 404 ) {
+          console.error("plugin settings not found");
+          settings = {
+            'Default' : {
+                auto_provision_resources : false,
+                device_response_data : '""'
+            }
+          };
+
+          startServer();
+
+        } else {
+          // Active wait until server is ready
+          console.error('server not available, checking again in 15 seconds...');
+          setTimeout( launchServer, 15000 );
+        }
+    });
+}
+
+launchServer();