+++ jupyter-minimal/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## 3.10.5-1 - 2022-08-23
+## 3.10.9-1 - 2023-03-23
 ### Added
 - First publication of plugin